Replace the closure system with interfaces.
Before, if one wanted to use the Populate()
function with any sort of complexity, they had to
use closures to keep variables persistent between
function calls. This system was confusing and
difficult to explain. Now, one only has to
provide an interface that implements a
GetFrame() function to Populate(), which I feel
is more intuitive and cleaner overall.